# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tasks application",
        "version": "777"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login",
                "description": "Login",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "request": "Login request",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "<access token>"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "data": {
                                            "properties": {
                                                "errors": {
                                                    "properties": {
                                                        "email": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string",
                                                                "example": "Email is invalid"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Fail login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "Incorrect email or password"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register",
                "description": "Register user",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "requestBody": {
                    "request": "Register request",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "<access token>"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "data": {
                                            "properties": {
                                                "errors": {
                                                    "properties": {
                                                        "email": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string",
                                                                "example": "Email is invalid"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tasks/list/{page}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Tasks list",
                "description": "Return list of tasks",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Page of list",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "per_page": {
                                                    "type": "int",
                                                    "example": "10"
                                                },
                                                "page": {
                                                    "type": "int",
                                                    "example": "1"
                                                },
                                                "count": {
                                                    "type": "int",
                                                    "example": "100"
                                                },
                                                "list": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Task"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tasks/create": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create task",
                "description": "Create task",
                "operationId": "9cbf95971b47b3009c243d3294cfb1a0",
                "requestBody": {
                    "request": "Task create request",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Saved result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tasks/delete/{id}": {
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete task",
                "description": "Delete task",
                "operationId": "40647bc8bfecc2875b790b7379f3b80e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tasks/update/{id}": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Update task",
                "description": "Update task",
                "operationId": "eccd7d8cd96011f3b929c0b7783a379d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "request": "Task update request",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "description": "Login request",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "user@email.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "LoginRequest"
                }
            },
            "RegisterRequest": {
                "description": "Register request",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "username"
                    },
                    "email": {
                        "type": "string",
                        "example": "user@email.com"
                    },
                    "password": {
                        "type": "password",
                        "example": "password"
                    },
                    "confirm": {
                        "type": "password",
                        "example": "password"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "RegisterRequest"
                }
            },
            "TaskCreateRequest": {
                "description": "Create task request",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "task 1"
                    },
                    "description": {
                        "type": "string",
                        "example": "description"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TaskCreateRequest"
                }
            },
            "TaskUpdateRequest": {
                "description": "Update task request",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "task 1"
                    },
                    "description": {
                        "type": "string",
                        "example": "description"
                    },
                    "completed": {
                        "type": "bool",
                        "example": "1"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TaskUpdateRequest"
                }
            },
            "Task": {
                "description": "Task model",
                "properties": {
                    "id": {
                        "type": "int",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "task 1"
                    },
                    "description": {
                        "type": "string",
                        "example": "description 1"
                    },
                    "completed": {
                        "type": "bool",
                        "example": "true"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Task"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}